feat: add logged out state and refactor filter into it's own component
Ref #182
